[feat] 설정 페이지 기본 틀 추가
